Move more stuff out of App.js
